python/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.110.0
Flask==3.0.2
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymango==0.1.1
pymongo==4.6.2
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
Werkzeug==3.0.1
